fix value iteration
Turns out the algorithm I had implemented was the same number of
iterations as value iteration, but not the same. each iteration of
Policy Evaluation invovles weigthing expected reward with how a policy Pi would choose actions.
value iteration simply assumes that the policy will be greedy with respect to V,
so it doesn't need to weight over pi.

resulting V is exactly equal to policy iteration! cooool!